requirements.txt
psycopg2-binary
aiohttp
requests
tqdm